Rename directory_file_mngr.c, directory_file.c symbols
