+++ asm/include/overlay_10.inc
@@ -3,7 +3,7 @@
 .public _020B0A5C
 .public _s32_div_f
 .public abs
-.public AllocAndLoadFileInPack
+.public DirectoryFileMngr_OpenDirectoryFile
 .public BAG_ITEMS_PTR_MIRROR
 .public CalcMenuHeightDiv8__020329E4
 .public CleanStickyItemsInBag
@@ -168,7 +168,7 @@
 .public GetDungeonGroup
 .public GetDungeonMode
 .public GetDungeonModeSpecial
-.public GetFileLengthInPackWithPackNb
+.public DirectoryFileMngr_GetDirectoryFileSize
 .public GetGameMode
 .public GetItemAtIdx
 .public GetMoneyCarried
